dev-rust/unicode-xid: 0.1.0 Use PN/PV in SRC_URI
As I mentioned in dev-rust/serde's commit message, this being missing
leads to risky annoying problems if people copy/bump the ebuild...
well, that happend.

This ebuild was specifying to use *serde* sources for a mere
*dependency* of serde, and fortunately, it worked for a different
reason:

The person who generated the Manifest clearly pulled
unicode-xid-0.1.0.crate from a gentoo distfiles mirror, which was
already the correct value, and so it generated it from that crate,
not the (wrong) one specified here in SRC_URI

That was a lucky bullet dodge. Phew!

Package-Manager: Portage-2.3.75, Repoman-2.3.16
Signed-off-by: Kent Fredric <kentnl@gentoo.org>