rust-util/pulldown-cmark: Add prototype for binary packages w/ v=0.7.0
This is a prototype design for splitting the binary targets of rust
crates into isolated packages, to allow concern seperating.

Especially, this allows us to restrict the use of USE flags in dev-rust
to mechanisms for communicating dependencies for --feature's, without
also forcing us to use the same namespace for selected features of the
binary.

Implementation wise, this works by faking a brand new crate with only
the binary aspects, and depending on the dev-rust/ library as-is, which
works nicely as a side effect of the cargo design having the same level
of isolation between binary and library crates.

The code is presently a bit rough and messy due to the current design
of rust-crate.eclass, which we need to crib some of to do our work.

The process of patching/stripping the original source tree also needs a
bit of love.

And a little thought needs to go into how we install documentation, if
any, and where it should go, as due to limitations in portage's design,
if we adopt official locations for both:

  dev-rust/pulldown-cmark-0.7.0
  rust-util/pulldown-cmark-0.7.0

Which, for both, is:

  /usr/share/doc/pulldown-cmark-0.7.0/

We would guarantee some collisions :/

Fortunately, dev-rust/ eschews proper use of this, and any documentation
is shipped inside the "crate" tree, which helps us dodge a sizeable
bullet.

Also, I've adopted a layout of:

  /usr/lib/rust/{binaries,examples}/${PN}/${PVR}/

Which guarantees infinite concurrent slotting, and also simplifies
 alternatives_auto_makesym 's ability to match the latest version,
without needing extended bash patterns and fragile regex to disambiguate
horror shows like:

  /usr/lib/rust/binaries/sha-utils-1.0.0
  /usr/lib/rust/binaries/sha-0.4.0

( Because a simple "sha-*" will of course match both, and building an
  expr that matches only one of the two, but also all versions of the
  one, has proven to have bugs )

Package-Manager: Portage-2.3.96, Repoman-2.3.22
Signed-off-by: Kent Fredric <kentnl@gentoo.org>